Moved BinaryStringFormatter into sensor package and made it package private
